+++ resources/views/components/tools/filters/livewire-component-filter.blade.php
@@ -1,0 +1,1 @@
+<livewire:dynamic-component :key="$wirekey" :is=$livewireComponent wire:model="{{ $modelableValue }}" />