+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to `laravel-livewire-tables` will be documented in this file
 - Add two classes for Bootstrap popover filter - "llt-bs4-filterpopover-mw" and "llt-bs5-filterpopover-mw", sets a min-width of 15rem
 - Add setSearchFieldAttributes() and getSearchFieldAttributes()
 - Add setFilterPopoverAttributes() and getFilterPopoverAttributes()
+- Create LivewireComponentFilter
 
 ## [Unreleased] - 3.x (beta-0)
 - Requirements Change